% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{compute_gvals}
\alias{compute_gvals}
\title{Compute the g values as defined in the thesis}
\usage{
compute_gvals(model, orig_data, nu = 1)
}
\arguments{
\item{model}{Neural Network for binary classification}

\item{orig_data}{The observed data.}

\item{nu}{Defaults to 1. Fraction of noise samples to real samples that
was used during training of the neural network.}
}
\value{
A vector containing the log-likelihood ratios between the NCE model and the
simplified vine
}
\description{
Compute the g values as defined in the thesis
}
