% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{compute_integral}
\alias{compute_integral}
\title{Monte Carlo integral of nce_cop (using importance sampling)}
\usage{
compute_integral(
  model,
  fitted_vine,
  n_samples,
  nu,
  data_dim_if_unif = -1,
  user_info = FALSE
)
}
\arguments{
\item{model}{The classifier trained to distinguish non-simplified data
with labels 1 from simplified data with labels 0}

\item{fitted_vine}{The vine copula fitted to the observed data.}

\item{n_samples}{Number of samples to create to compute the integral}

\item{nu}{Ratio of noise to true samples}

\item{data_dim_if_unif}{Data dimensionality. If this is given (and !=-1),
then the samples are drawn from a uniform distribution on [0,1]^d.
Otherwise the samples are drawn from the simplified vine}

\item{user_info}{Whether to display an information, which steps are currently running.}
}
\value{
The Monte Carlo integral approximation.
}
\description{
Monte Carlo integral of nce_cop (using importance sampling)
}
