% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{copula_pairs_ggplot}
\alias{copula_pairs_ggplot}
\title{Function with similar functionality as the pairs_copula_data function of
the rvinecopulib package. Plots histograms of data on the diagonal,
scatter plots on the upper triangle and
marginally normalized contours on the lower triangle.}
\usage{
copula_pairs_ggplot(
  data,
  uscale = TRUE,
  method = c("ecdf", "kde1d"),
  title = NULL
)
}
\arguments{
\item{data}{Data to plot}

\item{uscale}{True, if the data is already on the copula scale, False otherwise.}

\item{method}{Method to use for transforming the data to copula scale, if uscale is False.}

\item{title}{Title for the Plot.}
}
\value{
An object, which can be plotted using ggplotify::as.ggplot(object)
}
\description{
Function with similar functionality as the pairs_copula_data function of
the rvinecopulib package. Plots histograms of data on the diagonal,
scatter plots on the upper triangle and
marginally normalized contours on the lower triangle.
}
