% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{correction_factors}
\alias{correction_factors}
\title{Calculate the correction factors, i.e. the factors lambda, with which the
simplified copula needs to be multiplied to get the NCE copula.}
\usage{
correction_factors(model, obs, nu = 1, verbose = 1)
}
\arguments{
\item{model}{The classifier trained to distinguish non-simplified data
with labels 1 from simplified data with labels 0}

\item{obs}{the observations, for which the factors should be created.}

\item{nu}{The fraction of noise to real samples}

\item{verbose}{Whether output should be printed. Can be 0,1 or 2.}
}
\value{
The correction factors
}
\description{
Calculate the correction factors, i.e. the factors lambda, with which the
simplified copula needs to be multiplied to get the NCE copula.
}
