% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{count_NN_params}
\alias{count_NN_params}
\title{Count number of neural network parameters}
\usage{
count_NN_params(weights)
}
\arguments{
\item{weights}{The weights of a neural network, accessible via model$weights
for a tensorflow neural network called model}
}
\value{
num_params (integer): total number of parameters in the network
}
\description{
Count number of neural network parameters
}
