% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{fisher_z_transform}
\alias{fisher_z_transform}
\title{Computes the fisher Z transform (artanh)}
\usage{
fisher_z_transform(tau)
}
\arguments{
\item{tau}{the value of Kendall's tau to transform to a real number.
Can also be a vector of taus.}
}
\value{
The fisher transformation(s) of tau. If tau is a number, this returns a number
if tau was a vector this returns a vector
}
\description{
Computes the fisher Z transform (artanh)
}
