% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{get_max_matrix}
\alias{get_max_matrix}
\title{function to compute the matrix tilde(M)}
\usage{
get_max_matrix(mat)
}
\arguments{
\item{mat}{The input matrix}
}
\value{
the matrix, with entries tilde(m)_k,i = max_j=1,...,k (m_j,i)
}
\description{
function to compute the matrix tilde(M)
}
