% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{inverse_fisher_transform}
\alias{inverse_fisher_transform}
\title{Computes the inverse of the fisher Z transform, which is the tanh}
\usage{
inverse_fisher_transform(T_val)
}
\arguments{
\item{T_val}{The T-value that is supposed to be turned into a Kendall's tau again}
}
\value{
The inverse fisher transformation(s) of T_val. If T is a number, this returns a number
if T is a vector this returns a vector
}
\description{
Computes the inverse of the fisher Z transform, which is the tanh
}
