% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{log_likelihood_non_simp_parallel}
\alias{log_likelihood_non_simp_parallel}
\title{parallel evaluation of log likelihood data}
\usage{
log_likelihood_non_simp_parallel(
  u_data = matrix(c(0.2, 0.3, 0.4, 0.5, 0.5, 0.5), ncol = 3, byrow = TRUE),
  struct = matrix(c(1, 1, 1, 2, 2, 0, 3, 0, 0), byrow = TRUE, ncol = 3),
  families = list(list("gumbel", "gumbel"), list("gumbel")),
  params = list(c(2), c(1.3)),
  param_cond_funcs = list(list(u_to_param)),
  rotations = list(list(0, 0), list(0)),
  return_vector = FALSE
)
}
\arguments{
\item{u_data}{matrix of datapoints. Each row should by one datapoint, for which
a likelihood should be evaluated}

\item{struct}{rvine-structure matrix}

\item{families}{copula families, formatted as a list of lists}

\item{params}{parameters for the first tree, formatted as a list (of vectors)}

\item{param_cond_funcs}{list of list of functions that maps the u values, which the respective
copula is conditioned on, and a copula family-name to a parameter for that copula}

\item{rotations}{rotations of the copulas, formatted as a list of lists.}

\item{return_vector}{Boolean, defaults to FALSE. If TRUE, for every row
in u_data the function returns the likelihood for that row, i.e. returns a vector. Otherwise
it returns a single number (the sum of the individual log likelihoods)}
}
\value{
L: if return_vector is TRUE a vector, otherwise a number.
The log likelihoods of the individual samples, or of the whole dataset
}
\description{
parallel evaluation of log likelihood data
}
