% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{lr_schedule_fun}
\alias{lr_schedule_fun}
\title{learning rate scheduler: halves the learning rate every 30 epochs
argument to a keras function}
\usage{
lr_schedule_fun(epoch, lr)
}
\arguments{
\item{epoch}{current epoch}

\item{lr}{current learning rate}
}
\value{
lr: the updated learning rate
}
\description{
learning rate scheduler: halves the learning rate every 30 epochs
argument to a keras function
}
