% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{NCE_cop}
\alias{NCE_cop}
\title{compute the value of the NCE copula obtained from the simplified fit,
together with the classifier}
\usage{
NCE_cop(model, fitted_vine, obs, nu = 1, verbose = 1)
}
\arguments{
\item{model}{The classifier trained to distinguish non-simplified data
with labels 1 from simplified data with labels 0}

\item{fitted_vine}{The vine copula fitted to the observed data.}

\item{obs}{The observations, for which the NCE copula should be created.}

\item{nu}{The fraction of noise to real samples}

\item{verbose}{Whether output should be printed. Can be 0,1 or 2.}
}
\value{
The likelihoods of the copula derived via noise contrastive estimation
evaluated at the points given by obs.
}
\description{
compute the value of the NCE copula obtained from the simplified fit,
together with the classifier
}
