% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{perform_linear_quant_reg}
\alias{perform_linear_quant_reg}
\title{performs a linear quantile regression}
\usage{
perform_linear_quant_reg(
  covariates_train,
  covariates_test,
  response_train,
  response_test,
  bottom_quantiles_lin = seq(0.05, 0.25, 0.05),
  top_quantiles_lin = seq(0.75, 0.95, 0.05),
  method = "br"
)
}
\arguments{
\item{covariates_train}{matrix or dataframe. The covariates to train the quantile regression method.}

\item{covariates_test}{matrix or dataframe. The covariates to test the quantile regression method.}

\item{response_train}{Vector. Values to predict in the quantile regression (train set)}

\item{response_test}{Vector. Values to predict in the quantile regression (test set)}

\item{bottom_quantiles_lin}{Defaults to c(0.05,0.1). Vector of lower quantiles,
for which it is checked, whether the conditioned quantile estimates are >0.
Tests whether the alternative model is better}

\item{top_quantiles_lin}{Defaults to c(0.05,0.1). Vector of upper quantiles,
for which it is checked, whether the conditioned quantile estimates are <0.
Tests whether the simplified model is better}

\item{method}{Defaults to "br". Which method to use for the linear quantile regression}
}
\value{
: A List of 4 vectors, the first containing the number of samples, for which
q(bottom_quantile_levels) > 0 holds, the second containing the number of samples, for
which q(top_quantile_levels) < 0 holds. Behind that train and test pinball loss are listed.
}
\description{
performs a linear quantile regression
}
