% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{perform_quant_reg}
\alias{perform_quant_reg}
\title{performs a D-Vine quantile regression}
\usage{
perform_quant_reg(
  covariates_train,
  covariates_test,
  response_train,
  response_test,
  family_set_name = "nonparametric",
  bottom_quantile_levels = seq(0.05, 0.25, 0.05),
  top_quantile_levels = seq(0.75, 0.95, 0.05),
  cores = 1
)
}
\arguments{
\item{covariates_train}{matrix or dataframe. The covariates to train the quantile regression method.}

\item{covariates_test}{matrix or dataframe. The covariates to test the quantile regression method.}

\item{response_train}{Vector. Values to predict in the quantile regression (train set)}

\item{response_test}{Vector. Values to predict in the quantile regression (test set)}

\item{family_set_name}{Defaults to "nonparametric".
Argument family_set in the D-vine Quantile regression function.
Possible choices include "onepar" and "parametric".}

\item{bottom_quantile_levels}{Defaults to seq(0.05,0.25,0.05). Vector of lower quantiles,
for which it is checked, whether the conditioned quantile estimates are >0.
Tests whether the alternative model is better}

\item{top_quantile_levels}{Defaults to seq(0.75,0.95,0.05). Vector of upper quantiles,
for which it is checked, whether the conditioned quantile estimates are <0.
Tests whether the simplified model is better}

\item{cores}{How many cores to use for fitting the D-Vine quantile regression model}
}
\value{
: A List of 2 vectors, the first containing the number of samples, for which
q(bottom_quantile_levels) > 0 holds, the second containing the number of samples, for
which q(top_quantile_levels) < 0 holds.
}
\description{
performs a D-Vine quantile regression
}
