% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{permute_back}
\alias{permute_back}
\title{From the samples of the permuted structure, this function returns the samples from the original structure}
\usage{
permute_back(simulations, original_order)
}
\arguments{
\item{simulations}{Matrix with ncol(simulations) = length(original_order)
The samples from the reordered rvine structure (with antidiagonal values ordered
from bottom left to top right)}

\item{original_order}{The elements on the antidiagonal from bottom left to top right in the original structure}
}
\value{
result: the simulations, with the correct order.
}
\description{
From the samples of the permuted structure, this function returns the samples from the original structure
}
