% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{permute_forward}
\alias{permute_forward}
\title{From data coming from an rvine structure with elements current_order on the top right to
bottom left of the antidiagonal, this function returns the samples of the permuted structure,
where the elements on the antidiagonal are ordered increasingly from top right to bottom left.}
\usage{
permute_forward(samples, current_order)
}
\arguments{
\item{samples}{Matrix with ncol(simulations) = length(current_order)
The given data.}

\item{current_order}{The elements on the antidiagonal from bottom left to top right
in the original structure}
}
\value{
result: the data permuted to fit an order, that is increasing from the top
right to the bottom left of the r vine structure matrix.
}
\description{
From data coming from an rvine structure with elements current_order on the top right to
bottom left of the antidiagonal, this function returns the samples of the permuted structure,
where the elements on the antidiagonal are ordered increasingly from top right to bottom left.
}
