% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{permute_indices}
\alias{permute_indices}
\title{permutes indices of a rvine structure matrix so that on the antidiagonal
the values 1 to d (dimension of the matrix) are ordered from top right to bottom left}
\usage{
permute_indices(mat)
}
\arguments{
\item{mat}{The rvine structure matrix}
}
\value{
list with matrix as first entry, another list as second entry and a vector as third entry
First entry of result is the new structure matrix, second works as a dictionary, mapping the values
the last contains the values on the antidiagonal of the original matrix.
}
\description{
permutes indices of a rvine structure matrix so that on the antidiagonal
the values 1 to d (dimension of the matrix) are ordered from top right to bottom left
}
