% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Calculate the pinball loss (sometimes called check loss)}
\usage{
pinball_loss(y_true, y_pred, tau)
}
\arguments{
\item{y_true}{observed values}

\item{y_pred}{predicted quantiles}

\item{tau}{quantile level (in the thesis this is alpha,
tau for consistency with the other packages)}
}
\value{
A number, which is the pinball loss for the given quantile level
alpha, true values y_true and predicted values y_pred
}
\description{
Calculate the pinball loss (sometimes called check loss)
}
