% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_cond_to_ktau_1d}
\alias{plot_cond_to_ktau_1d}
\title{Plot of the cond to ktau function with 1 conditioning variable}
\usage{
plot_cond_to_ktau_1d(func_list, titles = -1, u_cond_vals = 1:99/100)
}
\arguments{
\item{func_list}{list of functions for which plots should be created}

\item{titles}{vector of titles for the plots corresponding to the different
functions in func_list}

\item{u_cond_vals}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values, for which the functions should be evaluated for the plot.}
}
\description{
Plot of the cond to ktau function with 1 conditioning variable
}
