% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_cond_to_ktau_2d}
\alias{plot_cond_to_ktau_2d}
\title{3d plot of the cond to ktau function, for 2 conditioning variables}
\usage{
plot_cond_to_ktau_2d(
  func_list,
  titles = -1,
  u_cond_vals_1 = 1:99/100,
  u_cond_vals_2 = 1:99/100
)
}
\arguments{
\item{func_list}{list of functions for which plots should be created}

\item{titles}{vector of titles for the plots corresponding to the different
functions in func_list}

\item{u_cond_vals_1}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values of the first conditioning variable, for which the
functions should be evaluated for the plot.}

\item{u_cond_vals_2}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values of the second conditioning variable, for which the
functions should be evaluated for the plot.}
}
\description{
3d plot of the cond to ktau function, for 2 conditioning variables
}
