% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_contours_1d}
\alias{plot_contours_1d}
\title{Marginally normalized contour plots for copulas with different
families and one-dimensional conditioning values}
\usage{
plot_contours_1d(
  cond_vals = c(0.2, 0.4, 0.6, 0.8),
  family_vals = c("frank", "gaussian", "gumbel", "clayton"),
  param_cond_func_1d = u_to_param_linear(c(1)),
  automatic_title = TRUE,
  manual_title = "",
  manual_subtitle = ""
)
}
\arguments{
\item{cond_vals}{The conditioning values
(plotted next to each other in columns)}

\item{family_vals}{The names of the copula families in the rows}

\item{param_cond_func_1d}{The parameter function to use}

\item{automatic_title}{Whether to automatically create a title (TRUE or FALSE)}

\item{manual_title}{Manually enter a title if desired. If this is not "",
then automatic_title is automatically set to FALSE}

\item{manual_subtitle}{Manually enter a subtitle if desired. If this is "",
then an automatic subtitle is set,in
if either automatic_title is TRUE or manual_title is not "".}
}
\description{
Marginally normalized contour plots for copulas with different
families and one-dimensional conditioning values
}
