% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_contours_2d}
\alias{plot_contours_2d}
\title{Marginally normalized contour plots for a copula with family family_name
with different two-dimensional conditioning values}
\usage{
plot_contours_2d(
  cond_vals_1 = c(0.2, 0.4, 0.6, 0.8),
  cond_vals_2 = c(0.2, 0.4, 0.6, 0.8),
  family_name = "frank",
  param_cond_func_2d = u_to_param_linear(c(0.7, 0.3)),
  automatic_title = TRUE,
  manual_title = "",
  manual_subtitle = ""
)
}
\arguments{
\item{cond_vals_1}{The conditioning values of the first variable
(plotted next to each other in columns)}

\item{cond_vals_2}{The conditioning values of the second variable
(plotted in the rows)}

\item{family_name}{The name of the copula family to use}

\item{param_cond_func_2d}{The parameter function to use}

\item{automatic_title}{Whether an automatically created title
should be plotted.Needs to be TRUE or FALSE.}

\item{manual_title}{Manual Title, if desired. If this is not "", then
automatic_title is automatically set to FALSE.}

\item{manual_subtitle}{Manual Subtitle, if desired. If this is "",
then an automatic subtitle is set,
if either automatic_title is TRUE or manual_title is not "".}
}
\description{
Marginally normalized contour plots for a copula with family family_name
with different two-dimensional conditioning values
}
