% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_ktau_contour_3d}
\alias{plot_ktau_contour_3d}
\title{Contour plot of ktau for 3 conditioning variables, always one fixed.}
\usage{
plot_ktau_contour_3d(
  cond_to_ktau_func,
  title = "",
  u_cond_vals_1 = 1:99/100,
  u_cond_vals_2 = 1:99/100,
  u_cond_vals_fixed = c(0.25, 0.5, 0.7)
)
}
\arguments{
\item{cond_to_ktau_func}{Function mapping conditioning values to kendall's tau values}

\item{title}{Title for the plot}

\item{u_cond_vals_1}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values of the first conditioning variable that is not fixed,
for which the functions should be evaluated for the plot.}

\item{u_cond_vals_2}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values of the second conditioning variable that is not fixed,
for which the functions should be evaluated for the plot.}

\item{u_cond_vals_fixed}{Vector, defaults to c(0.25,0.5,0.75). Contains the
conditioning values of the fixed variable.}
}
\description{
Contour plot of ktau for 3 conditioning variables, always one fixed.
}
