% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{qreg_train_test_split}
\alias{qreg_train_test_split}
\title{Train Test split for the data used in quantile regression}
\usage{
qreg_train_test_split(covariates, predictors, train_proportion)
}
\arguments{
\item{covariates}{matrix. The covariates, used to derive conditional quantiles}

\item{predictors}{vector or matrix. The values for which quantiles should be estimated}

\item{train_proportion}{Number, defaults to 0.9. Fraction of the data to use for training}
}
\value{
list with 4 elements: covariates_train, covariates_test, predictors_train, predictors_test
(i.e. original covariates and predictors split up randomly)
}
\description{
Train Test split for the data used in quantile regression
}
