% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{scaled_tanh}
\alias{scaled_tanh}
\title{Returns scaling_factor * tanh(T_val) + shift}
\usage{
scaled_tanh(T_val, scaling_factor = 1, shift = 0)
}
\arguments{
\item{T_val}{The T-value that is supposed to be turned into a Kendall's tau again}

\item{scaling_factor}{scale the result by this factor if desired. By default,
tanh returns values between -1 and 1. if a scaling_factor is specified, the result is
between - scaling factor and scaling factor}

\item{shift}{shift the result if desired}
}
\value{
The inverse fisher transformation(s) of T_val. If T is a number, this returns a number
if T is a vector this returns a vector
}
\description{
Returns scaling_factor * tanh(T_val) + shift
}
