% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{simulate_non_simp_parallel}
\alias{simulate_non_simp_parallel}
\title{parallel simulation of a non_simplified vine copula.}
\usage{
simulate_non_simp_parallel(
  n_samples = 500,
  struct = matrix(c(1, 1, 1, 2, 2, 0, 3, 0, 0), byrow = TRUE, ncol = 3),
  families = list(list("gumbel", "gumbel"), list("gumbel")),
  params = list(c(2), c(1.3)),
  param_cond_funcs = list(list(u_to_param)),
  rotations = list(list(0, 0), list(0))
)
}
\arguments{
\item{n_samples}{number of samples to simulate}

\item{struct}{rvine-structure matrix}

\item{families}{copula families, formatted as a list of lists}

\item{params}{parameters for the first tree, formatted as a list (of vectors)}

\item{param_cond_funcs}{list of list of functions that maps the u values, which the respective
copula is conditioned on, and a copula family-name to a parameter for that copula}

\item{rotations}{rotations of the copulas, formatted as a list of lists.}
}
\value{
result: A nxd matrix, where each row contains one sample.
}
\description{
parallel simulation of a non_simplified vine copula.
}
