% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{simulate_single_sample}
\alias{simulate_single_sample}
\title{Simulates a single sample,
This function is called by simulate_non_simplified, after modifying the input structure,
to be able to parallelize the procedure more easily for better performance}
\usage{
simulate_single_sample(
  new_struct,
  M_mat,
  families,
  params,
  param_cond_funcs,
  rotations
)
}
\arguments{
\item{new_struct}{rvine structure, where the antidiagonal is ordered
increasingly from top right to bottom left.}

\item{M_mat}{the matrix \eqn{\tilde{M}}, with \eqn{\tilde{m}_{k,i} = max_{j=1}^{k}{new_struct_{j,i}}}}

\item{families}{copula families, formatted as a list of lists}

\item{params}{parameters for the first tree, formatted as a list (of vectors)}

\item{param_cond_funcs}{list of list of functions that maps the u values, which the respective
copula is conditioned on, and a copula family-name to a parameter for that copula}

\item{rotations}{rotations of the copulas, formatted as a list of lists.}
}
\value{
u_row: a d-dimensional vector sampled from the non-simplified vine copula
}
\description{
Simulates a single sample,
This function is called by simulate_non_simplified, after modifying the input structure,
to be able to parallelize the procedure more easily for better performance
}
