% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{train_test_split}
\alias{train_test_split}
\title{Performs a train test split with original data already split in train and test
Gives all datapoints in orig_data a label of 1, all datapoints in simplified data a label of 0}
\usage{
train_test_split(orig_data_train, orig_data_test, simplified_data)
}
\arguments{
\item{orig_data_train}{The training set of the real data, which gets labels of 1}

\item{orig_data_test}{The test set of the real data, which gets labels of 1}

\item{simplified_data}{Synthetic data set, simulated from a simplified
vine copula, gets labels of 0}
}
\value{
A list with 4 entries, x_train, x_test, y_train, y_test in that order.
}
\description{
Performs a train test split with original data already split in train and test
Gives all datapoints in orig_data a label of 1, all datapoints in simplified data a label of 0
}
