% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_methods.R
\name{train_test_split_orig}
\alias{train_test_split_orig}
\title{Performs a train test split for the original data}
\usage{
train_test_split_orig(data, train_perc)
}
\arguments{
\item{data}{The data set}

\item{train_perc}{Defaults to 0.8. Percentage to use for training.}
}
\value{
A list with 2 entries: orig_train, orig_test
}
\description{
Performs a train test split for the original data
}
