% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{u_to_ktau_linear}
\alias{u_to_ktau_linear}
\title{Takes a vector a and returns a function of u (vector of elements between 0 and 1).
That function takes the dot product between a and u, scales that to
T_values and calculates a kendall's tau value using the tanh.}
\usage{
u_to_ktau_linear(a, tau_lower = -0.92, tau_upper = 0.92)
}
\arguments{
\item{a}{a vector which needs to be a convex combination
(i.e. all entries >=0 and they have to sum to 1)}

\item{tau_lower}{Lowest tau value, defaults to -0.92
(should be between -1 and 1 and less than tau_upper)}

\item{tau_upper}{Highest tau value, defaults to 0.92
(should be between -1 and 1 and greater than tau_lower)}
}
\value{
A function of u, which calculates the kendalls tau value of
a copula given the conditioned values.
}
\description{
Takes a vector a and returns a function of u (vector of elements between 0 and 1).
That function takes the dot product between a and u, scales that to
T_values and calculates a kendall's tau value using the tanh.
}
