% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{u_to_ktau_quadratic}
\alias{u_to_ktau_quadratic}
\title{Takes a vector a and returns a function of u (vector of elements between 0 and 1)
and of a string "family", which corresponds to a copula family.
that function transforms the values u using qnorm (inverse normal dist.function),
builds polynomial terms of degree 2 of these terms and then transforms them
to the interval (tau_lower, tau_upper) using a scaled tanh.
If dim(u) <=3 includes all terms u[i]*u[j], (powers and mixed terms).
If dim(u) > 3 only includes the individual powers, u[i]^k, k=1,2.}
\usage{
u_to_ktau_quadratic(a, tau_lower = -0.92, tau_upper = 0.92)
}
\arguments{
\item{a}{a vector of weights}

\item{tau_lower}{Lowest tau value
(should be between -1 and 1 and less than tau_upper)}

\item{tau_upper}{Highest tau value
(should be between -1 and 1 and greater than tau_lower)}
}
\value{
A function of u and family, which calculates the parameter of
a copula given the conditioned values.
}
\description{
Takes a vector a and returns a function of u (vector of elements between 0 and 1)
and of a string "family", which corresponds to a copula family.
that function transforms the values u using qnorm (inverse normal dist.function),
builds polynomial terms of degree 2 of these terms and then transforms them
to the interval (tau_lower, tau_upper) using a scaled tanh.
If dim(u) <=3 includes all terms u[i]*u[j], (powers and mixed terms).
If dim(u) > 3 only includes the individual powers, u[i]^k, k=1,2.
}
