% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_simplified_vine.R
\name{u_to_param_linear}
\alias{u_to_param_linear}
\title{takes a vector a and returns a function of u (vector of elements between 0 and 1)
and of a string "family", which corresponds to a copula family.
that function takes the dot product between a and u
and returns a linear function of that dot product.}
\usage{
u_to_param_linear(a, tau_lower = -0.92, tau_upper = 0.92)
}
\arguments{
\item{a}{a vector which needs to be a convex combination
(i.e. all entries >=0 and they have to sum to 1)}

\item{tau_lower}{Lowest tau value, defaults to -0.92
(should be between -1 and 1 and less than tau_upper)}

\item{tau_upper}{Highest tau value, defaults to 0.92
(should be between -1 and 1 and greater than tau_lower)}
}
\value{
A function of u and family, which calculates the parameter of
a copula given the conditioned values.
}
\description{
takes a vector a and returns a function of u (vector of elements between 0 and 1)
and of a string "family", which corresponds to a copula family.
that function takes the dot product between a and u
and returns a linear function of that dot product.
}
