% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_cond_to_ktau_3d}
\alias{plot_cond_to_ktau_3d}
\title{Several 3d Plots for in total 3 conditioning variables,
always one is fixed for the plots.}
\usage{
plot_cond_to_ktau_3d(
  cond_to_ktau_func,
  title = "",
  u_cond_vals_1 = 1:99/100,
  u_cond_vals_2 = 1:99/100,
  u_cond_vals_fixed = c(0.3, 0.7)
)
}
\arguments{
\item{cond_to_ktau_func}{Function mapping conditioning values to kendall's tau values}

\item{title}{Title of the plot.}

\item{u_cond_vals_1}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values of the first conditioning variable that is not fixed,
for which the functions should be evaluated for the plot.}

\item{u_cond_vals_2}{Vector, defaults to c(0.01,...,0.99). Contains the
conditioning values of the second conditioning variable that is not fixed,
for which the functions should be evaluated for the plot.}

\item{u_cond_vals_fixed}{Vector, defaults to c(0.3,0.7). Contains the
conditioning values of the fixed variable.}
}
\description{
Several 3d Plots for in total 3 conditioning variables,
always one is fixed for the plots.
}
